+++ .changeset/fluffy-kids-fix.md
@@ -1,0 +1,5 @@
+---
+'next-use-contextual-routing': minor
+---
+
+feat: support Next.js v15
