+++ .changeset/three-panthers-learn.md
@@ -1,0 +1,5 @@
+---
+'next-use-contextual-routing': major
+---
+
+refactor: drop support for Next.js v13/v14
