+++ README.md
@@ -122,6 +122,7 @@ The path to return to to close contextual navigation.
 | v2.2.0                      | v9.X -> v12.X  |
 | v3.0.0                      | v13.X          |
 | v3.2.0 +                    | v13.X -> v14.X |
+| v4.0.0 +                    | v15.X          |
 
 ## Notes
