+++ package.json
@@ -55,8 +55,7 @@
     "es-check": "^7.2.1",
     "jest": "^29.3.1",
     "jest-environment-jsdom": "^29.3.1",
-    "next": "^14.2.4",
-    "next-13": "npm:next@^13.5.6",
+    "next": "^15.0.0",
     "prettier": "^3.3.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -67,7 +66,7 @@
     "typescript": "^5.5.3"
   },
   "peerDependencies": {
-    "next": "^13.0.0 | ^14.0.0",
+    "next": "^15.0.0",
     "react": ">=18.0.0"
   },
   "jest": {
